% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCoMoMo.R
\name{fitCoMoMo}
\alias{fitCoMoMo}
\title{fitting different methods.}
\usage{
fitCoMoMo(
  models,
  data = NULL,
  Dxt = NULL,
  Ext = NULL,
  ages.fit = NULL,
  years.fit = NULL,
  ages = NULL,
  years = NULL
)
}
\description{
fitting different methods.
}
