% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Assigning names to different specified models.}
\usage{
model(models, modelNames)
}
\value{
A list of named mortality models.
}
\description{
Assigning names to different specified models.
}
