% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.CoMoMo}
\alias{plot.CoMoMo}
\title{Plot the weights from different model combination methods}
\usage{
\method{plot}{CoMoMo}(object, ...)
}
\arguments{
\item{object}{an object containing the weights for different horizons.

 @examples

 weights <- bma(modlist, data = DataStMoMo, ages.fit = agesFit, years.fit = yearsFit, h = 5, method = "cv")
 plot(weights)}
}
\description{
Plot the weights from different model combination methods
}
