% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoMoMo.R
\name{fitCoMoMo}
\alias{fitCoMoMo}
\title{Model combination using different methods.}
\usage{
fitCoMoMo(
  models,
  data = NULL,
  Dxt = NULL,
  Ext = NULL,
  ages.fit = NULL,
  years.fit = NULL,
  ages = NULL,
  years = NULL
)
}
\description{
Model combination using different methods.
}
