% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{valloss}
\alias{valloss}
\title{Weights estimation using different model combinations methods.}
\usage{
valloss(
  models,
  data = NULL,
  Dxt = NULL,
  Ext = NULL,
  ages.fit = NULL,
  years.fit = NULL,
  ages = NULL,
  years = NULL,
  holdout = NULL,
  h = NULL
)
}
\arguments{
\item{models}{are the specified models to be combined.}

\item{data}{an optional object of type StMoMoData containing

information on deaths and exposures to be used for training the model.

This is typically created with function \code{\link{StMoMoData}}.

If this is not provided then the training data is taken from

arguments, \code{Dxt}, \code{Ext}, \code{ages}, \code{years}.}

\item{Dxt}{optional matrix of deaths data.}

\item{Ext}{optional matrix of observed exposures of the same
dimension of \code{Dxt}.}

\item{ages.fit}{optional vector of ages to include in the
training. Must be a subset of \code{ages}.}

\item{years.fit}{optional vector of years to include in the
training. Must be a subset of \code{years}.}

\item{ages}{optional vector of ages corresponding to rows of
\code{Dxt} and \code{Ext}.}

\item{years}{optional vector of years corresponding to rows of
\code{Dxt} and \code{Ext}.}

\item{h}{number of years for forecasting horizon.}
}
\description{
We consider four different model combination approaches:
 Simple Model Averaging, Bayesian Model Averaging, Model Confidence Set,
 and Stacked Regression Ensemble. These methods vary from each other depending on
 how they use the historical data to choose the combining weights or the models to be combined.
}
\details{
Simple Model Averaging assigns equal weights to all the models.

Bayesian Model Averaging estimates the weights using the posterior model probabilities.

Model Confidence Set chooses the subset of superior mortality models to combine
where each model is assigned equal weight

Stacked Regression Ensemble combines point forecasts from
multiple base learners using the weights that optimize a
cross-validation criterion.
}
